add dataset

+++ PMFI/src/main/java/pmfi/functions/IProbabilistic.java
@@ -2,6 +2,6 @@ package pmfi.functions;
 
 public interface IProbabilistic{
     int calculateProbabilisticSupport(double minProbabilisticConfidence);
-    boolean isProbabilisticFrequentItemset(int minSupport, double minProbabilisticConfidence);
-    boolean isProbabilisticMaximalFrequentItemset(int minSupport, double minProbabilisticConfidence);
+    boolean isProbabilisticFrequentItemset(double minSupport, double minProbabilisticConfidence);
+    boolean isProbabilisticMaximalFrequentItemset(double minSupport, double minProbabilisticConfidence);
 }
